sort/serializers.py
from rest_framework.serializers import ModelSerializer
from sort.models import Category, Tag

class CategorySerializer(ModelSerializer):
    class Meta:
        model = Category
        fields = ('id', 'name', 'link')

class TagSerializer(ModelSerializer):
    class Meta:
        model = Tag
        fields = ('id', 'name', 'link')